# lib/api-zod/src/generated/types/inventoryState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Part } from "./part";

export interface InventoryState {
  parts: Part[];
  headers: string[];
  announcement: string;
  showAnnouncement: boolean;
}