# lib/api-zod/src/generated/types/partInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PartInput {
  partNumber: string;
  name?: string;
  quantity?: string;
  location?: string;
  price?: string;
  images?: string[];
  [key: string]: unknown;
}